Moved group generation logic to another file
